# Store "marked" flag in high bit of "roots" counter