@mwerezak 2022-03-12
I can remove the `#[inline]`s if you think they are unnecessary.